Move check_direct_build to uv-build-backend
Move check_direct_build to uv-build-backend and add a parameter for logging. When using  this function as part of the resolver or installer, the log messages will otherwise be lacking context.
