+++ crates/uv-build-backend/Cargo.toml
@@ -20,6 +20,7 @@ uv-normalize = { workspace = true }
 uv-pep440 = { workspace = true }
 uv-pep508 = { workspace = true }
 uv-pypi-types = { workspace = true }
+uv-version = { workspace = true }
 uv-warnings = { workspace = true }
 
 csv = { workspace = true }
