+++ crates/uv/src/commands/build_frontend.rs
@@ -7,7 +7,7 @@ use std::{fmt, io};
 use anyhow::{Context, Result};
 use owo_colors::OwoColorize;
 use thiserror::Error;
-use tracing::{debug, instrument, trace};
+use tracing::instrument;
 
 use crate::commands::pip::operations;
 use crate::commands::project::find_requires_python;
@@ -16,7 +16,7 @@ use crate::commands::ExitStatus;
 use crate::printer::Printer;
 use crate::settings::{ResolverSettings, ResolverSettingsRef};
 use uv_auth::store_credentials;
-use uv_build_backend::PyProjectToml;
+use uv_build_backend::check_direct_build;
 use uv_cache::{Cache, CacheBucket};
 use uv_client::{BaseClientBuilder, Connectivity, FlatIndexClient, RegistryClientBuilder};
 use uv_configuration::{
@@ -581,13 +581,16 @@ async fn build_package(
             return Err(Error::ListForcePep517);
         }
 
-        if !check_direct_build(source.path()) {
+        if !check_direct_build(source.path(), source.path().user_display()) {
             // TODO(konsti): Provide more context on what mismatched
             return Err(Error::ListNonUv);
         }
 
         BuildAction::List
-    } else if preview.is_enabled() && !force_pep517 && check_direct_build(source.path()) {
+    } else if preview.is_enabled()
+        && !force_pep517
+        && check_direct_build(source.path(), source.path().user_display())
+    {
         BuildAction::DirectBuild
     } else {
         BuildAction::Pep517
@@ -1204,37 +1207,5 @@ impl BuildPlan {
                 }
             }
         })
-    }
-}
-
-/// Check if the build backend is matching the currently running uv version.
-fn check_direct_build(source_tree: &Path) -> bool {
-    let pyproject_toml: PyProjectToml =
-        match fs_err::read_to_string(source_tree.join("pyproject.toml"))
-            .map_err(anyhow::Error::from)
-            .and_then(|pyproject_toml| Ok(toml::from_str(&pyproject_toml)?))
-        {
-            Ok(pyproject_toml) => pyproject_toml,
-            Err(err) => {
-                debug!("Not using uv build backend direct build, no pyproject.toml: {err}");
-                return false;
-            }
-        };
-    match pyproject_toml
-        .check_build_system(uv_version::version())
-        .as_slice()
-    {
-        // No warnings -> match
-        [] => true,
-        // Any warning -> no match
-        [first, others @ ..] => {
-            debug!(
-                "Not using uv build backend direct build, pyproject.toml does not match: {first}"
-            );
-            for other in others {
-                trace!("Further uv build backend direct build mismatch: {other}");
-            }
-            false
-        }
     }
 }
